Refactor: 글자or색 맞추기 게임에서 중복되는 코드를 묶은 커스텀 훅 추가
