+++ src/pages/games/ShownColor.tsx
@@ -1,10 +1,8 @@
-import { useEffect, useState } from 'react';
 import { GameProps } from '../../routes/gameRouter';
-import { useGameLogic } from '../../hooks/useGameLogic';
 import GameQuestion from '../../components/common/GameQuestion';
 import { Button, ButtonContainer } from '../../components/common/GameButton';
 import { Container, Letter } from '../../components/games/ShownColor';
-import { getRandomFloat } from '../../utils/random';
+import { useColorGameLogic } from '../../hooks/useColorGameLogic';
 
 function ShownColor({
   gameData,
@@ -14,40 +12,18 @@ function ShownColor({
   setAnswerState,
   answerState,
 }: GameProps) {
-  const [candidates, setCandidates] = useState<string[]>([]);
-  const [letter, setLetter] = useState('');
-  const [color, setColor] = useState('');
-  const { onClickButton, setAnswer, buttonRefs } = useGameLogic<string>({
-    gameData,
-    onGameEnd,
-    saveGameResult,
-    isNextButtonClicked,
-    setAnswerState,
-    answerState,
-  });
-  const colorNames = ['빨강', '파랑', '노랑', '검정'];
-  const colors = ['red', 'blue', 'yellow', 'black'];
-  const colorNameMatch: { [key: string]: string } = {
-    red: '빨강',
-    blue: '파랑',
-    yellow: '노랑',
-    black: '검정',
-  };
-
-  useEffect(() => {
-    const randomIndex1 = Math.floor(getRandomFloat() * 3);
-    const name = colorNames[randomIndex1];
-    const newColors = colors.filter((_, i) => i !== randomIndex1);
-    const randomIndex2 = Math.floor(getRandomFloat() * 2);
-    const color = newColors[randomIndex2];
-
-    setAnswer(colorNameMatch[color]);
-    setLetter(name);
-    setColor(color);
-    setCandidates(
-      [name, colorNameMatch[color]].sort(() => getRandomFloat() - 0.5),
+  const { onClickButton, buttonRefs, candidates, letter, color } =
+    useColorGameLogic(
+      {
+        gameData,
+        onGameEnd,
+        saveGameResult,
+        isNextButtonClicked,
+        setAnswerState,
+        answerState,
+      },
+      false,
     );
-  }, []);
 
   return (
     <Container>
